Make lowering error tests less fragile
